+++ app/preprints/-components/preprint-metrics/styles.scss
@@ -1,0 +1,122 @@
+// stylelint-disable max-nesting-depth, selector-max-compound-selectors
+
+.file-renderer-container {
+    width: 100%;
+    height: 1150px;
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    justify-content: flex-start;
+
+    .file-container {
+        width: 100%;
+        height: 1100px;
+        display: flex;
+        flex-direction: column;
+        align-items: flex-start;
+        justify-content: flex-start;
+    }
+
+    .details-container {
+        width: 100%;
+        height: 50px;
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        justify-content: flex-start;
+
+        .name-container {
+            width: 60%;
+            height: 50px;
+            overflow: hidden;
+            display: flex;
+            flex-direction: row;
+            align-items: center;
+            justify-content: space-between;
+
+            .name,
+            .version {
+                display: flex;
+                flex-direction: row;
+                align-items: center;
+                justify-content: flex-start;
+                height: 25px;
+            }
+        }
+
+        .version-container {
+            width: 40%;
+            height: 50px;
+            display: flex;
+            flex-direction: row;
+            align-items: center;
+            justify-content: flex-end;
+
+            .version-menu {
+                background-color: $color-bg-white;
+                border: 1px solid rgba(0, 0, 0, 0.15);
+                border-radius: 4px;
+                box-shadow: 0 6px 12px rgba(0, 0, 0, 0.175);
+                padding: 5px;
+                margin: 2px 0 0;
+            
+                ul {
+                    list-style: none;
+                    padding-inline-start: 0;
+                }
+
+                .li-container {
+                    margin-top: 10px;
+                    margin-bottom: 10px;
+
+                    .btn {
+                        display: inline-block;
+                        margin-bottom: 0;
+                        font-weight: 400;
+                        text-align: center;
+                        white-space: nowrap;
+                        touch-action: manipulation;
+                        cursor: pointer;
+                        background-image: none;
+                        border: 1px solid transparent;
+                        border-radius: 2px;
+                        padding: 6px 12px;
+                        font-size: 14px;
+                        line-height: 1.42857;
+                        user-select: none;
+                        vertical-align: middle;
+                        color: $color-text-white;
+                    }
+
+                    .btn-primary {
+                        color: $color-text-white;
+                        background-color: $color-bg-blue-dark;
+                        border-color: #2e6da4;
+                    }
+                }
+            }
+        }
+    }
+
+    &.mobile {
+        height: 500px;
+
+        .file-container {
+            height: 400px;
+        }
+
+        .details-container {
+            height: 100px;
+            flex-direction: column;
+
+            .version-container,
+            .name-container {
+                height: 50px;
+                width: 100%;
+                justify-content: center;
+                align-items: flex-start;
+                flex-direction: column;
+            }
+        }
+    }
+}
